+++ database/migrations/2018_10_19_082942_add_forum_link_to_data_sets.php
@@ -1,0 +1,32 @@
+<?php
+
+use Illuminate\Support\Facades\Schema;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+class AddForumLinkToDataSets extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::table('data_sets', function (Blueprint $table) {
+            $table->string('forum_link', 191)->nullable();
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('data_sets', function (Blueprint $table) {
+            $table->dropColumn('forum_link');
+        });
+    }
+}